Remove duplication in History and Alarms
Duplicate code in History and Alarms was factored out
into a new class named RecordContainer. Now, Alarms is empty
but kept for semantic reasons for the time being.
