+++ lib/obix.rb
@@ -9,6 +9,7 @@ module OBIX
   autoload :Watch, "obix/watch"
   autoload :Alarms, "obix/alarms"
   autoload :History, "obix/history"
+  autoload :RecordContainer, "obix/record_container"
   autoload :Tag, "obix/tag"
   autoload :Builder, "obix/builder"
   autoload :Configuration, "obix/configuration"
